+++ crates/teloxide-core/src/local_macros.rs
@@ -1168,6 +1168,22 @@ macro_rules! requester_forward {
             $body!(get_business_connection this (business_connection_id: BusinessConnectionId))
         }
     };
+    (@method get_managed_bot_token $body:ident $ty:ident) => {
+        type GetManagedBotToken = $ty![GetManagedBotToken];
+
+        fn get_managed_bot_token(&self, user_id: UserId) -> Self::GetManagedBotToken {
+            let this = self;
+            $body!(get_managed_bot_token this (user_id: UserId))
+        }
+    };
+    (@method replace_managed_bot_token $body:ident $ty:ident) => {
+        type ReplaceManagedBotToken = $ty![ReplaceManagedBotToken];
+
+        fn replace_managed_bot_token(&self, user_id: UserId) -> Self::ReplaceManagedBotToken {
+            let this = self;
+            $body!(replace_managed_bot_token this (user_id: UserId))
+        }
+    };
     (@method get_my_commands $body:ident $ty:ident) => {
         type GetMyCommands = $ty![GetMyCommands];
 
@@ -1286,6 +1302,14 @@ macro_rules! requester_forward {
         fn save_prepared_inline_message(&self, user_id: UserId, result: InlineQueryResult) -> Self::SavePreparedInlineMessage {
             let this = self;
             $body!(save_prepared_inline_message this (user_id: UserId, result: InlineQueryResult))
+        }
+    };
+    (@method save_prepared_keyboard_button $body:ident $ty:ident) => {
+        type SavePreparedKeyboardButton = $ty![SavePreparedKeyboardButton];
+
+        fn save_prepared_keyboard_button(&self, user_id: UserId, button: KeyboardButton) -> Self::SavePreparedKeyboardButton {
+            let this = self;
+            $body!(save_prepared_keyboard_button this (user_id: UserId, button: KeyboardButton))
         }
     };
     (@method edit_message_text $body:ident $ty:ident) => {