+++ crates/teloxide-core/src/net/request.rs
@@ -213,7 +213,8 @@ mod tests {
                     "poll_answer":{
                         "poll_id":"POLL_ID",
                         "user": {"id":42,"is_bot":false,"first_name":"blah"},
-                        "option_ids": []
+                        "option_ids": [],
+                        "option_persistent_ids": []
                     }
                 }
             ]
@@ -235,7 +236,8 @@ mod tests {
                     "poll_answer":{
                         "poll_id":"POLL_ID",
                         "user": {"id":42,"is_bot":false,"first_name":"blah"},
-                        "option_ids": []
+                        "option_ids": [],
+                        "option_persistent_ids": []
                     }
                 },
                 {
@@ -247,7 +249,8 @@ mod tests {
                     "poll_answer":{
                         "poll_id":"POLL_ID",
                         "user": {"id":42,"is_bot":false,"first_name":"blah"},
-                        "option_ids": [3, 4, 8]
+                        "option_ids": [3, 4, 8],
+                        "option_persistent_ids": ["a", "b", "c"]
                     }
                 },
                 {