+++ crates/teloxide-core/src/types/chat_member.rs
@@ -743,6 +743,7 @@ mod tests {
                 is_premium: false,
                 has_topics_enabled: false,
                 allows_users_to_create_topics: false,
+                can_manage_bots: false,
                 added_to_attachment_menu: false,
             },
             kind: ChatMemberKind::Administrator(Administrator {
@@ -813,6 +814,7 @@ mod tests {
                 is_premium: false,
                 has_topics_enabled: false,
                 allows_users_to_create_topics: false,
+                can_manage_bots: false,
                 added_to_attachment_menu: false,
             },
             kind: ChatMemberKind::Restricted(Restricted {