+++ crates/teloxide-core/src/types/me.rs
@@ -78,6 +78,7 @@ mod tests {
                 is_premium: false,
                 has_topics_enabled: false,
                 allows_users_to_create_topics: false,
+                can_manage_bots: false,
                 added_to_attachment_menu: false,
             },
             can_join_groups: false,